modified constraint on PrimitiveGeoFeature.doubleProperty
git-svn-id: http://svn.codehaus.org/geoserver/trunk@8502 ef1d6e69-97e5-0310-af46-8a06194da32a

Former-commit-id: 9aa08cc24a7b6c5993eccd66c54b89ccaa52b564 [formerly 071db45e20940933ecb8759fd868ade355a88189]
Former-commit-id: 19ddc0aa37cd314bdb8c560c5b2122cf9efcf7db

+++ configuration/citewfs-1.1-h2/dataset-sf0.sql
@@ -12,7 +12,7 @@ ALTER TABLE "PrimitiveGeoFeature" ADD "uriProperty" varchar;
 ALTER TABLE "PrimitiveGeoFeature" ADD "measurand" float not null;
 ALTER TABLE "PrimitiveGeoFeature" ADD "dateTimeProperty" datetime;
 ALTER TABLE "PrimitiveGeoFeature" ADD "dateProperty" datetime;
-ALTER TABLE "PrimitiveGeoFeature" ADD "decimalProperty" decimal(4,2) not null;
+ALTER TABLE "PrimitiveGeoFeature" ADD "decimalProperty" double(2) not null;
 ALTER TABLE "PrimitiveGeoFeature" ADD "relatedFeature" varchar; 
 ALTER TABLE "PrimitiveGeoFeature" ADD "id" int AUTO_INCREMENT(1);
 ALTER TABLE "PrimitiveGeoFeature" ADD CONSTRAINT "surfacePropertyGeometryType" CHECK GeometryType("surfaceProperty") IS NULL OR GeometryType("surfaceProperty") = 'POLYGON';
